+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+## [1.5.0] - 2020-04-19
+### Fixed
+- [#322](https://gitlab.com/dctx/rapidpass/rapidpass-api/-/issues/332) - Implemented registrar user account 
+locking for several failed login attempts
 
 ## [1.4.8] - 2020-04-19
 ### Added